Fix: Separate React Query Provider into a Client Component
Query Provider를 루트 `layout.tsx`에 바로 사용하지 않고, 클라이언트 컴포넌트로 분리해서 에러 해결
- Theme Provider에 분리한 Query Provider를 합침
- 참고 블로그(https://velog.io/@ahhpc2012/Error-Only-plain-objects-and-a-few-built-ins-can-be-passed-to-Client-Components-from-Server-Components.-Classes-or-null-prototypes-are-not-supported)
