+++ apps/web/src/_shared/clientBoundary/ProviderClient.tsx
@@ -1,0 +1,24 @@
+'use client';
+
+import { ThemeProvider as NextThemesProvider } from 'next-themes';
+import { getQueryClient } from '@workspace/react-query/queryClient';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+
+const queryClient = getQueryClient();
+
+export const ProviderClient = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <NextThemesProvider
+      attribute="class"
+      defaultTheme="system"
+      enableSystem
+      disableTransitionOnChange
+      enableColorScheme>
+      <QueryClientProvider client={queryClient}>
+        {children}
+        <ReactQueryDevtools />
+      </QueryClientProvider>
+    </NextThemesProvider>
+  );
+};
