+++ apps/web/src/app/layout.tsx
@@ -1,11 +1,8 @@
 import '@workspace/ui/globals.css';
 
-import { ThemeProviderClient } from '@/_shared/clientBoundary/ThemeProviderClient';
-import { QueryClientProvider } from '@tanstack/react-query';
-import { getQueryClient } from '@workspace/react-query/queryClient';
-import type { Metadata } from 'next';
+import { ProviderClient } from '@/_shared/clientBoundary/ProviderClient';
 import localFont from 'next/font/local';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import type { Metadata } from 'next';
 
 const pretendard = localFont({
   src: '../../../../packages/ui/src/fonts/pretendard/PretendardVariable.woff2',
@@ -26,8 +23,6 @@ export const metadata: Metadata = {
   description: '우학동 : 귀찮았던 동아리 관리, 우학동이 대신 해 드릴게요',
 };
 
-const queryClient = getQueryClient();
-
 /**
  * 서버에서 렌더링된 HTML과 클라이언트에서의 초기 렌더링이 정확히 일치하지 않을 수 있는 경우, 이를 방지하기 위해 suppressHydrationWarning을 사용
  * @see https://ui.shadcn.com/docs/dark-mode/next
@@ -40,12 +35,7 @@ export default function RootLayout({
   return (
     <html lang="ko" suppressHydrationWarning>
       <body className={`${pretendard.variable} ${jua.variable} antialiased`}>
-        <ThemeProviderClient>
-          <QueryClientProvider client={queryClient}>
-            {children}
-            <ReactQueryDevtools />
-          </QueryClientProvider>
-        </ThemeProviderClient>
+        <ProviderClient>{children}</ProviderClient>
       </body>
     </html>
   );
